reorganize ScanData/Paths
consolidated ScanPaths and ScanData, etc. into a single project, rather than two separate ones. Modified all imports accordingly. Removed ImageAnalysis depence on ScanData
