--- GEECS-Data-Utils/Geecs-Paths-Utils/pyproject.toml
@@ -1,15 +1,0 @@
-[tool.poetry]
-name = "geecs_paths_utils"
-version = "0.0.1"
-description = ""
-authors = ["Sam Barber <sbarber@lbl.gov>"]
-readme = "README.md"
-packages = [{include = "geecs_paths_utils"}]
-
-
-[tool.poetry.dependencies]
-python = ">=3.7,<3.11"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
