--- GEECS-Data-Utils/Geecs-ScanData-Utils/README.md
@@ -1,1 +1,0 @@
-# Data Utils Module
