+++ GEECS-Data-Utils/geecs_data_utils/scan_data.py
@@ -5,10 +5,11 @@ import shutil
 from pathlib import Path
 from typing import Optional, Union
 
+import numpy as np
 import pandas as pd
+import nptdms as tdms
 
-from geecs_paths_utils import ScanPaths
-from geecs_scan_data_utils.utils import read_geecs_tdms
+from geecs_data_utils.scan_paths import ScanPaths
 
 # Module-level logger
 logger = logging.getLogger(__name__)
@@ -17,6 +18,43 @@ if not logging.getLogger().hasHandlers():
         level=logging.INFO,
         format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
     )
+
+def read_geecs_tdms(file_path: Path) -> Optional[dict[str, dict[str, np.ndarray]]]:
+    """
+    Read a GEECS TDMS file and return nested dict of device -> variable -> ndarray.
+    """
+    if not file_path.is_file() or file_path.suffix.lower() != '.tdms':
+        return None
+
+    with tdms.TdmsFile.open(str(file_path)) as f_tdms:
+        def convert(channel: tdms.TdmsChannel):
+            arr = channel[:]
+            try:
+                return arr.astype('float64')
+            except ValueError:
+                return arr
+
+        return {
+            group.name: {
+                var.name.split(group.name)[1].lstrip('_'): convert(var)
+                for var in group.channels()
+            }
+            for group in f_tdms.groups()
+        }
+
+def geecs_tdms_dict_to_panda(data_dict: dict[str, dict[str, np.ndarray]]) -> pd.DataFrame:
+    """
+    Convert nested TDMS dict into a multi-indexed pandas DataFrame.
+    """
+    return (
+        pd.concat(
+            map(pd.DataFrame, data_dict.values()),
+            keys=data_dict.keys(),
+            axis=1
+        )
+        .set_index('Shotnumber')
+    )
+
 
 class ScanData(ScanPaths):
     """GEECS experiment scan with scalar loading."""
@@ -71,10 +109,7 @@ class ScanData(ScanPaths):
             analysis_sfile.unlink()
         shutil.copy2(src=scan_sfile, dst=analysis_sfile)
 
-
-
 if __name__ == "__main__":
-    tag = ScanData.get_scan_tag(2025, 5, 1, number=10, experiment="Undulator")
+    tag = ScanData.get_scan_tag(2025, 5, 7, number=10, experiment="Undulator")
     sd = ScanData(tag=tag)
-    sd.load_scalar_data()
-    print(sd.data_frame)
+    print(sd.get_sfile_data())
