+++ GEECS-Data-Utils/geecs_data_utils/scan_paths.py
@@ -10,8 +10,9 @@ from pathlib import Path
 from datetime import datetime, date
 from configparser import ConfigParser, NoSectionError
 from typing import Optional, Union
-from geecs_paths_utils.utils import ScanTag,SysPath,ConfigurationError, month_to_int
-from geecs_paths_utils.geecs_paths_config import GeecsPathsConfig
+from geecs_data_utils.utils import ScanTag, SysPath, ConfigurationError, month_to_int
+from geecs_data_utils.geecs_paths_config import GeecsPathsConfig
+
 
 # module‐level logger
 logger = logging.getLogger(__name__)
@@ -24,7 +25,7 @@ if not logging.getLogger().hasHandlers():
 class ScanPaths:
     """ Represents a GEECS experiment scan """
 
-    paths_config: Optional[GeecsPathsConfig] = None
+    paths_config = None
 
     def __init__(self, folder: Optional[SysPath] = None,
                  tag: Optional[ScanTag] = None,
@@ -431,5 +432,6 @@ class ScanPaths:
 ScanPaths.reload_paths_config()
 
 if __name__ == '__main__':
+
     test_tag = ScanPaths.get_scan_tag(2025, 4, 3, number=2, experiment='Undulator')
     print(ScanPaths.get_scan_folder_path(test_tag))
