+++ GEECS-Data-Utils/pyproject.toml
@@ -1,10 +1,10 @@
 [tool.poetry]
-name = "geecs_scan_data_utils"
+name = "geecs_data_utils"
 version = "0.0.1"
 description = ""
 authors = ["Sam Barber <sbarber@lbl.gov>"]
 readme = "README.md"
-packages = [{include = "geecs_scan_data_utils"}]
+packages = [{include = "geecs_data_utils"}]
 
 
 [tool.poetry.dependencies]
@@ -12,10 +12,6 @@ python = ">=3.9,<3.11"
 
 pandas = "^2.0"
 nptdms = "^1.7.0"
-
-geecs_paths_utils = { path = "../GEECS-Paths-Utils", develop = true }
-
-
 
 [build-system]
 requires = ["poetry-core"]
