+++ GEECS-PythonAPI/geecs_python_api/analysis/scans/__init__.py
@@ -5,11 +5,11 @@ __all__ = ["ScanData"]
 def __getattr__(name):
     if name == "ScanData":
         warnings.warn(
-            "geecs_python_api.analysis.scans.scan_data.ScanData has moved to geecs_scan_data_utils.scan_data.ScanData; "
+            "geecs_python_api.analysis.scans.scan_data.ScanData has moved to geecs_data_utils.scan_data.ScanData; "
             "please update your imports.",
             DeprecationWarning,
             stacklevel=2,
         )
-        from geecs_scan_data_utils.scan_data import ScanData
+        from geecs_data_utils import ScanData
         return ScanData
     raise AttributeError(f"module {__name__} has no attribute {name}")
