+++ GEECS-PythonAPI/htu_scripts/calibrations/steering_calibration.py
@@ -9,7 +9,7 @@ from geecs_python_api.controls.devices.HTU.diagnostics.screens import Screen
 from geecs_python_api.controls.devices.HTU.diagnostics import EBeamDiagnostics
 from geecs_python_api.controls.devices.HTU.diagnostics.cameras import Camera
 from geecs_python_api.tools.interfaces.prompts import text_input
-from geecs_scan_data_utils.scan_data import ScanData
+from geecs_data_utils import ScanData
 import geecs_python_api.analysis.scans.scan_analysis as sa
