+++ GEECS-PythonAPI/pyproject.toml
@@ -31,9 +31,7 @@ mysql-connector-python = "^8.2.0"
 
 # internal libraries
 ImageAnalysis = { path = "../ImageAnalysis", develop = true }
-geecs_scan_data_utils = { path = "../GEECS-Data-Utils/GEECS-ScanData-Utils", develop = true }
-
-
+geecs_data_utils = { path = "../GEECS-Data-Utils", develop = true }
 
 [build-system]
 requires = ["poetry-core"]
