+++ GEECS-Scanner-GUI/geecs_scanner/utils/module_open_folder.py
@@ -6,7 +6,7 @@ import re
 import subprocess
 import configparser
 from pathlib import Path
-from geecs_scan_data_utils.scan_data import ScanData
+from geecs_data_utils import ScanData
 
 def open_folder(path_to_folder: Path):
     """ Opens Windows file explorer to the specified location """
