+++ ImageAnalysis/image_analysis/offline_analyzers/Undulator/ACaveMagCam3.py
@@ -5,9 +5,7 @@ from pathlib import Path
 
 import numpy as np
 from image_analysis.base import ImageAnalyzer
-from geecs_paths_utils.scan_paths import ScanPaths
-from geecs_paths_utils.utils import ScanTag
-
+from geecs_data_utils import ScanPaths, ScanTag
 
 class ACaveMagCam3ImageAnalyzer(ImageAnalyzer):
