+++ ImageAnalysis/pyproject.toml
@@ -22,8 +22,6 @@ pypng = "^0.20220715.0"
 imageio = "^2.31.1"
 scikit-image = ">=0.19,<1.0"
 
-geecs_paths_utils = {path = "../GEECS-Data-Utils/GEECS-Paths-Utils", develop = true}
-
 # other
 boto3 = "^1.28"
 opencv-python = "^4.9.0.80"
