+++ ScanAnalysis/multiscan_analysis/Undulator/rad2_spectrum_stitcher.py
@@ -1,4 +1,4 @@
-from geecs_scan_data_utils.scan_data import ScanData
+from geecs_data_utils import ScanData
 from scan_analysis.analyzers.Undulator.rad2_spec_analysis import Rad2SpecAnalysis
 
 import numpy as np
