+++ ScanAnalysis/scan_analysis/analyzers/Undulator/mag_spec_stitcher_analysis.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Union, Optional
 if TYPE_CHECKING:
-    from geecs_paths_utils.scan_paths import ScanTag
+    from geecs_data_utils import ScanTag
 import logging
 from pathlib import Path
 import pandas as pd
@@ -209,7 +209,7 @@ class MagSpecStitcherAnalysis(ScanAnalysis):
 
 
 if __name__ == "__main__":
-    from geecs_scan_data_utils.scan_data import ScanData
+    from geecs_data_utils import ScanData
     tag = ScanData.get_scan_tag(year=2025, month=2, day=27, number=4, experiment_name='Undulator')
     analyzer = MagSpecStitcherAnalysis(scan_tag=tag, device_name="U_BCaveMagSpec", skip_plt_show=True)
     analyzer.run_analysis()
