+++ ScanAnalysis/scan_analysis/mapping/map_Undulator.py
@@ -1,7 +1,4 @@
-from geecs_paths_utils.scan_paths import ScanTag
-from geecs_scan_data_utils.scan_data import ScanData
-
-from geecs_paths_utils.scan_paths import ScanTag
+from geecs_data_utils import ScanData, ScanTag
 
 from scan_analysis.base import AnalyzerInfo as Info
 from scan_analysis.analyzers.common.array2D_scan_analysis import Array2DScanAnalysis
