+++ tests/ScannerTools/test_scanner_tools/test_probe_phase_analysis.py
@@ -8,8 +8,7 @@ import pytest
 
 from pathlib import Path
 
-from geecs_scan_data_utils.scan_data import ScanData
-from geecs_paths_utils.utils import ScanTag
+from geecs_data_utils import ScanData,ScanTag
 from image_analysis.offline_analyzers.density_from_phase_analysis import PhaseAnalysisConfig, PhasePreprocessor, PhaseDownrampProcessor
 
 def get_path_to_phase_file():
